Add files via upload
As a part of the research I was doing into Caribbean migration patterns, I handled resettlement records of Cubans in the 1960s. I transferred this data to Excel and created a couple of Matlab programs to visualize the data utilizing startup files from my Engineering Computing class at the University of Notre Dame. CubanRegistration.m is the main file for graphing the data from CubanMapData.xlsx in line graphs. CubanMap.m is the main file for graphing the data from CubanResettlementfigures.xlsx in a map of the world as well as displaying a legend that shows ranges of resettlement. The rest of the files are used in the startup files from Chad A. Greene's map files.